* Common/ContigProperties.h (opt::k): Unsigned rather than int. * AdjList/AdjList.cpp: Ditto. * Graph/adjtodot.cc: Ditto. * Overlap/Overlap.cpp: Ditto. * ParseAligns/ParseAligns.cpp: Ditto. * PathOverlap/PathOverlap.cpp: Ditto. * PopBubbles/PopBubbles.cpp: Ditto. * SimpleGraph/SimpleGraph.cpp: Ditto.
git-svn-id: https://svn.bcgsc.ca/svn/abyss/trunk@61266 ad02020f-2a26-4133-947e-1829317c1416

+++ Overlap/Overlap.cpp
@@ -61,7 +61,7 @@ static const char USAGE_MESSAGE[] =
 "Report bugs to <" PACKAGE_BUGREPORT ">.\n";
 
 namespace opt {
-	int k; // used by ContigGraph
+	unsigned k; // used by ContigGraph
 	static unsigned minimum_overlap = 5;
 	static int mask = 1;
 	static int scaffold = 1;
@@ -177,7 +177,7 @@ static FastaRecord overlapContigs(const ContigNode& t_id,
 {
 	string t = sequence(t_id);
 	string h = sequence(h_id);
-	assert(overlap < (unsigned)opt::k - 1);
+	assert(overlap < opt::k - 1);
 	unsigned gap = opt::k - 1 - overlap;
 	string a(t, t.length() - opt::k+1, gap);
 	string o(h, 0, overlap);
